collector/requirements.txt
websocket-client
requests>=2.31.0
psycopg2-binary
pandas>=2.2.2
pyyaml
pycryptodome
pytz>=2024.1
websocket-client>=1.6.0
